@@ -1,18 +1,18 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-  </PropertyGroup>
+	<PropertyGroup>
+		<OutputType>Exe</OutputType>
+		<TargetFramework>net6.0</TargetFramework>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="rebus" Version="6.0.0" />
-    <PackageReference Include="serilog" Version="2.9.0" />
-    <PackageReference Include="serilog.sinks.coloredconsole" Version="3.0.1" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="rebus" Version="6.0.0" />
+		<PackageReference Include="serilog" Version="2.9.0" />
+		<PackageReference Include="serilog.sinks.coloredconsole" Version="3.0.1" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Rebus.Serilog\Rebus.Serilog.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\Rebus.Serilog\Rebus.Serilog.csproj" />
+	</ItemGroup>
 
 </Project>
